requirements.txt
asgiref==3.6.0
black==23.3.0
cachetools==5.3.0
certifi==2023.5.7
cfgv====3.3.1
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
Django==4.2.1
djangorestframework==3.14.0
filelock==3.12.0
flake8==6.0.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
identify==2.5.24
idna====3.4
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
nodeenv==1.7.0
packaging==23.1
pathspec==0.11.1
pip====23.1.2
platformdirs==3.5.0
pre-commit==3.3.1
protobuf==4.22.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
pytz====2023.3
PyYAML==6.0
requests==2.30.0
rsa====4.9
setuptools==67.6.1
six====1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.23.0